rutas/rutas_service/applications/api/authentication.py
# instituciones/authentication.py

import jwt
from rest_framework import authentication, exceptions
from django.conf import settings

class JWTAuthentication(authentication.BaseAuthentication):
    def authenticate(self, request):
        auth_header = authentication.get_authorization_header(request)
        if not auth_header:
            return None  # No se proporciona token

        try:
            prefix, token = auth_header.decode('utf-8').split(' ')
            if prefix.lower() != 'bearer':
                return None
        except ValueError:
            return None

        return self.authenticate_credentials(token)

    def authenticate_credentials(self, token):
        try:
            payload = jwt.decode(token, settings.SECRET_KEY, algorithms=['HS256'])
            user_id = payload['user_id']
        except jwt.ExpiredSignatureError:
            raise exceptions.AuthenticationFailed('El token ha expirado.')
        except jwt.InvalidTokenError:
            raise exceptions.AuthenticationFailed('Token inválido.')

        # Creamos un usuario anónimo con el user_id extraído
        user = SimpleUser(user_id)
        return (user, token)

class SimpleUser:
    def __init__(self, user_id):
        self.id = user_id
        self.is_authenticated = True
        self.is_staff = False
        self.is_superuser = False
